chore: remove remote restore feature
